requirements.txt
openai>=1.78.1    # Core OpenAI API
openai-agents>=0.0.14 # OpenAI Agents SDK
pydantic>=2.10,<3
requests==2.31.0
beautifulsoup4==4.12.3
pandas==2.1.4
python-dotenv==1.0.1
matplotlib==3.8.4
tabulate==0.9.0     # For Markdown table generation
